Revist how Scala syntax is enabled
