+++ core/src/test/scala/io/finch/syntax/ScalaMapperSyntaxSpec.scala
@@ -1,11 +1,11 @@
 package io.finch.syntax
 
-import io.finch.syntax.scalaFutures._
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
 class ScalaMapperSyntaxSpec extends MapperSyntaxSpec {
 
-  it should behave like endpointMapper[Future]
+  import scalaFutures._
 
+  it should behave like endpointMapper[Future]
 }
